requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
anyio==4.4.0
attrs==24.2.0
certifi==2024.8.30
discord.py==2.4.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.8
multidict==6.0.5
ollama==0.3.2
sniffio==1.3.1
yarl==1.9.7
